feat(defaultErrorHandler): Do not invoke default error handler for ABORTED transitions
BREAKING CHANGE: ABORTED transitions do not invoke the `defaultErrorHandler`

Returning `false` from a transition hook will abort the transition.

- #### Old behavior

Previously, this case was considered an error and was logged by
`defaultErrorHandler`.
After your feedback, we agree that this is not typically an error.

- #### New behavior

Now, aborted transitions do not trigger the `defaultErrorHandler`

- #### Motivation:

> Why introduce a BC?

Most users do not consider ABORT to be an error.  The default error
handler should match this assumption.

- #### BC liklihood

> How likely am I to be affected?

Low: Most users do not consider ABORT to be an error. For most users
this will not be a BC.

- #### BC severity

> How severe is this BC?

Low: Users who want to handle all transition rejections can
register a `.onError` handler and filter/process accordingly.